Move OpeningHour to shared components
