+++ src/scss/includes/panels/categories.scss
@@ -144,14 +144,6 @@ $category_panel_head_height : 40px;
   overflow: hidden;
 }
 
-.category__panel__item {
-  .poi_panel__info__hours__status__text {
-    margin: 0;
-    font-size: 14px;
-    color: #353c52;
-  }
-}
-
 .category__panel__status {
   color: #353c52;
   font-size: 14px;
