+++ src/scss/main.scss
@@ -35,3 +35,4 @@
 
 @import "includes/ui_components";
 @import "includes/reviewScore";
+@import "includes/openingHour";
